+++ transfers/util.py
@@ -29,13 +29,14 @@ import numpy as np
 from constants import SRID_WGS84, SRID_UTM_ZONE_13N
 from db import Thing, Location
 from services.gcs_helper import get_storage_bucket
-from services.lexicon_mapper import lexicon_mapper
+
+# from services.lexicon_mapper import lexicon_mapper
 from services.util import (
     transform_srid,
     get_epqs_elevation_from_point,
-    get_state_from_point,
-    get_county_from_point,
-    get_quad_name_from_point,
+    # get_state_from_point,
+    # get_county_from_point,
+    # get_quad_name_from_point,
 )
 from transfers.logger import logger
 
@@ -310,6 +311,65 @@ def timeit(func):
         return timeit_direct(func, *args, **kwargs)
 
     return wrapper
+
+
+class LexiconMapper:
+    def __init__(self):
+        self._mappers = None
+
+    def map_value(self, value):
+        return self._make_lu_to_lexicon_mapper().get(value, value)
+
+    def _make_lu_to_lexicon_mapper(self):
+        if self._mappers:
+            return self._mappers
+
+        lu_tables = [
+            # "LU_AltitudeDatum",     # the code is the value, so no need for mapping
+            "LU_AltitudeMethod",  # CODE/MEANING
+            "LU_CollectionMethod",  # CODE/MEANING
+            "LU_ConstructionMethod",  # CODE/MEANING
+            "LU_CoordinateAccuracy",  # CODE/MEANING
+            # "LU_CoordinateDatum",   # the code is the value, so no need for mapping
+            "LU_CoordinateMethod",  # CODE/MEANING
+            "LU_CurrentUse",  # CODE/MEANING
+            "LU_DataQuality",  # CODE/MEANING
+            "LU_DataSource",  # CODE/MEANING
+            "LU_Depth_CompletionSource",  # CODE/MEANING
+            "LU_Discharge_ChemistrySource",  # CODE/MEANING
+            # "LU_FieldNoteTypes",    # not being used in the transfers since there are no records
+            # "LU_Formations",        # needs to be cleaned before it can be used
+            "LU_LevelStatus",  # CODE/MEANING
+            # "LU_Lithology",         # needs to be cleaned before it can be used
+            "LU_MajorAnalyte",  # CODE/MEANING
+            "LU_MeasurementMethod",  # CODE/MEANING
+            # "LU_MeasuringAgency",   # the abreviation is what is used in the new schema
+            "LU_MinorTraceAnalyte",  # CODE/MEANING
+            "LU_MonitoringStatus",  # CODE/MEANING
+            "LU_SampleType",  # CODE/MEANING
+            "LU_SiteType",  # CODE/MEANING
+            "LU_Status",  # CODE/MEANING
+        ]
+
+        mappers = {}
+
+        for lu_table in lu_tables:
+            table = read_csv(lu_table)
+
+            for i, row in table.iterrows():
+                if lu_table == "LU_Formations":
+                    code = row.Code
+                    meaning = row.Meaning
+                else:
+                    code = row.CODE
+                    meaning = row.MEANING
+
+                mappers.update({f"{lu_table}:{code}": meaning})
+        self._mappers = mappers
+        return mappers
+
+
+lexicon_mapper = LexiconMapper()
 
 
 # ============= EOF =============================================
